Add option to toggle CJK line breaking rules.
This adds a feature to RichTextLabel nodes. When a RichTextLabel is
changed it now queries the "rich_text/force_cjk_linebreaks" setting. If
the setting exists AND is enabled then all rich text is treated as CJK
text. If the setting does not exist OR is not enabled then the
previously implemented behavior (CJK detection by scanning text) is
used.

"rich_text/force_cjk_linebreaks" is the default name for the setting.
However, the name of the setting can be changed at compile-time by
defining GNOMESORT_CJK_LINEBREAK_SETTING to your preferred string.

To support this, RichTextLabels now connect to the
ProjectSettings::on_settings_changed signal. When the signal is emitted,
RichTextLabels will invalidate their current formatting and update. This
forces the CJK line break setting to be queried again. A method is added
to RichTextLabel called _settings_changed to facilitate this.

Obviously, projects don't have this setting by default so it will need
to be added. Once "rich_text/force_cjk_linebreaks" (or an equivalent)
has been added to the project it can be toggled from GDScript.